# laravel/task_Agiletech/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/by_region": {
            "get": {
                "tags": [
                    "by_region"
                ],
                "summary": "Get a count of brands by_region",
                "operationId": "85a664fc62c7be803f1a7ddc859a936e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/by_district": {
            "get": {
                "tags": [
                    "by_district"
                ],
                "summary": "Get a count of brands by_district",
                "operationId": "d87321d1faa6fdbb931e09c498698957",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "summary": "Register a new user",
                "operationId": "c0354e671ce642c7b4ff996bc4e475a0",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "User's phone_number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "401": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Authenticate user and generate JWT token",
                "operationId": "dde0e6a0694ae97f804fba7d40433c68",
                "parameters": [
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "User's phone_number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Unauthorised. Invalid credentials"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "summary": "Get logout user.",
                "operationId": "6ed3b558c7c74e809518c164c4e3a1a6",
                "responses": {
                    "200": {
                        "description": "Yuo are logout."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get a list of users",
                "operationId": "971db9bc974c0e607288cfe7ecc10096",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/user/": {
            "post": {
                "summary": "Store a newl user created resource in storage.",
                "operationId": "5db1ab81dbc69e2e6d6b811a3d5f35ac",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "user's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "user's phone_number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "user's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The user was successfully created"
                    }
                }
            }
        },
        "/api/user/id": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Display the details of user.",
                "operationId": "04f65273b5ef0aaf11360a7e9a768b6b",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update the  user in storage.",
                "operationId": "1dcf0396be6e572fb79829fab1b1d158",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "user's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "user's phone_number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "user's passwords",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The user was successfully updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Remove the user from storage.",
                "operationId": "d61875f395fe5ac1f803329915d28c6e",
                "responses": {
                    "200": {
                        "description": "The user was successfully deleted"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}